Refactor upload_test and download_test to not need internal lfs access & use NewRepo
